Fix misleading errmsg when driver doesn't support Chrome version
The old message ("Unable to automate Chrome version because it's too
old for this version of Chromedriver") assumes a situation. I, on the
other hand, encountered a situation where the version of the Chrome
browser was higher than the versions the Chromedriver could handle.

So instead of assuming things, a regexp was added to extract the
supported Chrome versions by Chromedriver, and display them on a
separate line.