fix(postinstall): avoid potential infinite loop
Resolves appium/appium#17196

If the postinstall script (`install-npm.js`) is unable to load the install script, the script will now fail instead of infinitely retrying to load the module.  The script was rather complicated, so I simplified it (as well as the build workflow).

The `install-npm.js` script will spawn an `npm run build` if the project doesn't yet exist (which is only applicable in a dev environment).

Fixed a couple broken things in `package.json` as well.